Implement transforms (formerly processors)
This commit adds support for single and chained transforms, enabling a
variety of fun and interesting things like markdown parsing.

Currently, both the input and output of a transform must be valid utf8,
but I'm planning to allow transforms to output arbitrary byte streams in
the future. Unfortunately, allowing non-utf8 _input_ is a lot harder
because Tera only supports utf8 and there's no simple way to get Tera to
ignore a file.