Overhaul github page
Added a custom layout, stylesheet, and font. Set up some templating
stuff as well, and a Makefile to simplify repetitive commands.

Still should rewrite the index page at some point to not be a duplicate
of the readme. Maybe I'll turn it into a shell script that outputs the
contents of the readme instead, to reduce duplication, or maybe I'll
turn this into more of a marketing/getting-started type thing. Idk.