Refactor site build logic
The site-building logic was pretty gross, and it was also massively
bloating the `cli::build` module which is not really the right place for
it.

This commit extracts all that logic to a new `site` module. It also
organizes the logic in a more object-oriented style, with the page and
layout Tera engines constructed ahead of time and stored in a `Site`
struct along with the config, which is now only loaded once. The actual
build logic is moved to a separate `Site::build` method, which uses
those template engines.

The contents of the `build` method are still mostly unchanged and still
need more refactoring. One thing that was changed, however, is that the
entire config struct is now inserted as the rendering context.